+++ src/main/java/pe/upc/finanzas/service/IAdministradorService.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import pe.upc.finanzas.entity.Administrador;
 
-public interface IAdministradorService extends CrudService<Administrador, Integer>{
+public interface IAdministradorService extends CrudService<Administrador, Long>{
 
 	Optional<Administrador> findByNAdministrador(String NAdministrador) throws Exception;
 }
