+++ src/main/java/pe/upc/finanzas/service/IPeriodo_TasaService.java
@@ -2,6 +2,6 @@ package pe.upc.finanzas.service;
 
 import pe.upc.finanzas.entity.Periodo_Tasa;
 
-public interface IPeriodo_TasaService extends CrudService<Periodo_Tasa,Integer> {
+public interface IPeriodo_TasaService extends CrudService<Periodo_Tasa,Long> {
 
 }
