+++ src/main/java/pe/upc/finanzas/service/ITipoMantenimientoService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import pe.upc.finanzas.entity.TipoMantenimiento;
 
-public interface ITipoMantenimientoService extends CrudService<TipoMantenimiento, Integer>{
+public interface ITipoMantenimientoService extends CrudService<TipoMantenimiento, Long>{
 
 	List<TipoMantenimiento> findByNTipoMantenimiento(String NTipoMantenimiento);
