+++ src/main/java/pe/upc/finanzas/service/ITipo_DeliveryService.java
@@ -2,6 +2,6 @@ package pe.upc.finanzas.service;
 
 import pe.upc.finanzas.entity.Tipo_Delivery;
 
-public interface ITipo_DeliveryService extends CrudService<Tipo_Delivery, Integer> {
+public interface ITipo_DeliveryService extends CrudService<Tipo_Delivery, Long> {
 
 }
